plugins/ckn_inference_daemon/requirements.txt
requests~=2.32.3
neo4j~=5.22.0
confluent-kafka~=2.4.0
pandas~=2.2.2
werkzeug~=3.1.3
torch~=2.6.0
pillow~=10.4.0
torchvision~=0.21.0
uvicorn~=0.34.0
fastapi~=0.115.8
kafka-python~=2.0.2
huggingface-hub~=0.28.1
python-multipart